3.0

+++ app/src/main/java/org/d3if3112/db/FoodEntity.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 data class FoodEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val foodName: String,
-    val calories: Int
+    val calories: Double
 )
